requirements.txt
Flask
yfinance
ta
python-dotenv
flask-socketio
scipy
numpy
pytest
pandas